% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPFilter.R
\name{apply_csp_filters}
\alias{apply_csp_filters}
\title{Apply per-band CSP filters to each trial and concatenate outputs.}
\usage{
apply_csp_filters(data, filters)
}
\arguments{
\item{data}{List of trial matrices.}

\item{filters}{List of projection matrices, typically one per band.}
}
\value{
List of projected trial matrices.
}
\description{
Apply per-band CSP filters to each trial and concatenate outputs.
}
