% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ATM.R
\name{atmBinTimeBinary}
\alias{atmBinTimeBinary}
\title{Time-bin binary activity matrix by OR aggregation.}
\usage{
atmBinTimeBinary(B, bin = 1L)
}
\arguments{
\item{B}{Binary activity matrix (\verb{time x channels}).}

\item{bin}{Bin width in time steps.}
}
\value{
Binned binary matrix.
}
\description{
Time-bin binary activity matrix by OR aggregation.
}
