% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ATM.R
\name{atmBranchingRatio}
\alias{atmBranchingRatio}
\title{Estimate branching ratio from avalanche segments.}
\usage{
atmBranchingRatio(B, segs)
}
\arguments{
\item{B}{Binary activity matrix.}

\item{segs}{Avalanche segments.}
}
\value{
Ratio of next-step to current-step active nodes, or NA if undefined.
}
\description{
Estimate branching ratio from avalanche segments.
}
