% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ATM.R
\name{atmChooseBin}
\alias{atmChooseBin}
\title{Choose the bin size whose branching ratio is closest to criticality.}
\usage{
atmChooseBin(
  Z,
  zThresh = 3,
  useAbs = TRUE,
  minLen = 2L,
  binCandidates = c(1L, 2L, 3L)
)
}
\arguments{
\item{Z}{Z-scored matrix.}

\item{zThresh}{Threshold used to binarize activity.}

\item{useAbs}{Whether thresholding is absolute (\code{abs(Z)}).}

\item{minLen}{Minimum avalanche length.}

\item{binCandidates}{Candidate bin widths.}
}
\value{
Selected bin width.
}
\description{
Choose the bin size whose branching ratio is closest to criticality.
}
