% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ATM.R
\name{atmCounts}
\alias{atmCounts}
\title{Count channel transitions within avalanche segments.}
\usage{
atmCounts(B, segs)
}
\arguments{
\item{B}{Binary activity matrix.}

\item{segs}{Avalanche segments from \code{atmFindAvalanches()}.}
}
\value{
List with pair counts \code{countIJ} and source counts \code{countI}.
}
\description{
Count channel transitions within avalanche segments.
}
