% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ATM.R
\name{atmFindAvalanches}
\alias{atmFindAvalanches}
\title{Detect contiguous avalanche segments.}
\usage{
atmFindAvalanches(B, minLen = 2L)
}
\arguments{
\item{B}{Binary activity matrix.}

\item{minLen}{Minimum segment length to keep.}
}
\value{
List of \verb{[start, end]} index pairs.
}
\description{
Detect contiguous avalanche segments.
}
