% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ATM.R
\name{atmProbMatrix}
\alias{atmProbMatrix}
\title{Convert transition counts to row-normalized transition probabilities.}
\usage{
atmProbMatrix(countIJ, countI)
}
\arguments{
\item{countIJ}{Transition count matrix.}

\item{countI}{Source activation counts.}
}
\value{
Transition probability matrix.
}
\description{
Convert transition counts to row-normalized transition probabilities.
}
