% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ATM.R
\name{atmTest}
\alias{atmTest}
\title{Transform test trials to ATM features using a fitted ATM object.}
\usage{
atmTest(x, object)
}
\arguments{
\item{x}{Trial list or single trial matrix.}

\item{object}{Fitted ATM object from \code{atmTrain()}.}
}
\value{
Feature matrix.
}
\description{
Transform test trials to ATM features using a fitted ATM object.
}
