% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ATM.R
\name{atmZscoreApply}
\alias{atmZscoreApply}
\title{Apply pre-fitted z-score normalization to one trial matrix.}
\usage{
atmZscoreApply(M, zfit)
}
\arguments{
\item{M}{Trial matrix.}

\item{zfit}{Output object from \code{atmZscoreFit()}.}
}
\value{
Z-scored matrix.
}
\description{
Apply pre-fitted z-score normalization to one trial matrix.
}
