% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ATM.R
\name{atmZscoreFit}
\alias{atmZscoreFit}
\title{Fit channel-wise z-score parameters across training trials.}
\usage{
atmZscoreFit(trialList)
}
\arguments{
\item{trialList}{List of trial matrices.}
}
\value{
List with channel means (\code{mu}) and standard deviations (\code{sd}).
}
\description{
Fit channel-wise z-score parameters across training trials.
}
