% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ACM.R, R/aliases.R
\name{compute_ACM}
\alias{compute_ACM}
\alias{compute_acm}
\title{Compute augmented covariance matrices from delayed embeddings.}
\usage{
compute_ACM(x, order = 2, delay = 1, shrinkage = c("no", "LW", "oas"))

compute_acm(x, order = 2, delay = 1, shrinkage = c("no", "LW", "oas"))
}
\arguments{
\item{x}{Trial matrix or list of trial matrices (\verb{samples x channels}).}

\item{order}{Embedding order (number of delayed blocks).}

\item{delay}{Delay between blocks in samples.}

\item{shrinkage}{Covariance estimator: \code{no}, \code{LW}, or \code{oas}.}
}
\value{
SPD covariance matrix or list of matrices matching input shape.
}
\description{
Compute augmented covariance matrices from delayed embeddings.
}
\examples{
x <- lapply(1:4, function(i) matrix(rnorm(200 * 3), 200, 3))
acm <- compute_ACM(x, order = 2, delay = 1, shrinkage = "oas")
dim(acm[[1]])
}
