% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtraction.R
\name{.apply_trial_preprocessor}
\alias{.apply_trial_preprocessor}
\title{Apply a fitted preprocessor to each trial matrix.}
\usage{
.apply_trial_preprocessor(trials, preproc)
}
\arguments{
\item{trials}{List of trial matrices.}

\item{preproc}{Object returned by \code{.fit_trial_preprocessor()}.}
}
\value{
List of transformed trial matrices.
}
\description{
Apply a fitted preprocessor to each trial matrix.
}
