% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtraction.R
\name{.resolve_cov_fun}
\alias{.resolve_cov_fun}
\title{Resolve covariance estimator function from type string.}
\usage{
.resolve_cov_fun(cov_type)
}
\arguments{
\item{cov_type}{One of \code{"oas"} or \code{"lw"}.}
}
\value{
Covariance estimator function.
}
\description{
Resolve covariance estimator function from type string.
}
