% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Riemannian_utils.R
\name{exp_map}
\alias{exp_map}
\title{Apply the affine-invariant exponential map from tangent to SPD space.}
\usage{
exp_map(P_omega, S_i)
}
\arguments{
\item{P_omega}{Reference SPD matrix or list with \code{sqrt} and \code{inv_sqrt}.}

\item{S_i}{Tangent matrix at \code{P_omega}.}
}
\value{
SPD matrix on the manifold.
}
\description{
Apply the affine-invariant exponential map from tangent to SPD space.
}
