% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{mdm_predict}
\alias{mdm_predict}
\title{Predict class labels with a trained MDM classifier.}
\usage{
mdm_predict(model, x)
}
\arguments{
\item{model}{An \code{"mdm"} object from \code{\link[=mdm_train]{mdm_train()}}.}

\item{x}{Trial list or a single trial matrix.}
}
\value{
A factor of predicted classes (levels match the training classes).
The trial-by-class Riemannian distance matrix is attached as attribute
\code{"distances"}.
}
\description{
Predict class labels with a trained MDM classifier.
}
\seealso{
\code{\link[=mdm_train]{mdm_train()}}
}
