% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{predict.mdm}
\alias{predict.mdm}
\title{Predict method for MDM classifiers.}
\usage{
\method{predict}{mdm}(object, x, type = c("class", "distance"), ...)
}
\arguments{
\item{object}{An \code{"mdm"} object from \code{\link[=mdm_train]{mdm_train()}}.}

\item{x}{Trial list or a single trial matrix.}

\item{type}{\code{"class"} for the predicted-class factor (default) or
\code{"distance"} for the trial-by-class distance matrix.}

\item{...}{Unused.}
}
\value{
A factor of class labels, or the distance matrix when
\code{type = "distance"}.
}
\description{
Predict method for MDM classifiers.
}
\seealso{
\code{\link[=mdm_train]{mdm_train()}}, \code{\link[=mdm_predict]{mdm_predict()}}
}
