% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msvar.R
\name{predict.msvar}
\alias{predict.msvar}
\title{Predict method for MSVAR classifiers.}
\usage{
\method{predict}{msvar}(object, x, type = c("class", "loglik"), ...)
}
\arguments{
\item{object}{An \code{"msvar"} object from \code{\link[=msvar_train]{msvar_train()}}.}

\item{x}{Trial list or a single trial matrix.}

\item{type}{\code{"class"} for the predicted-class factor (default) or
\code{"loglik"} for the trial-by-class log-likelihood matrix.}

\item{...}{Unused.}
}
\value{
A factor of class labels, or the log-likelihood matrix when
\code{type = "loglik"}.
}
\description{
Predict method for MSVAR classifiers.
}
\seealso{
\code{\link[=msvar_train]{msvar_train()}}, \code{\link[=msvar_predict]{msvar_predict()}}
}
