% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSPFilter.R
\name{prepare_csp_filters}
\alias{prepare_csp_filters}
\title{Train and package CSP-family filters for train/test pipelines.}
\usage{
prepare_csp_filters(
  data_train,
  labels_train,
  fs,
  channels,
  csp_type,
  frequency_bands = list(c(8, 30), c(11, 20), c(21, 30), c(31, 40)),
  ncomps = c(3, 3)
)
}
\arguments{
\item{data_train}{List of training trials.}

\item{labels_train}{Training labels.}

\item{fs}{Sampling rate.}

\item{channels}{Optional subset of channels.}

\item{csp_type}{One of \code{FBCSP} or \code{FBCSSP}.}

\item{frequency_bands}{List of band ranges.}

\item{ncomps}{Components per stage.}
}
\value{
List with projected training patterns and trained filters.
}
\description{
Train and package CSP-family filters for train/test pipelines.
}
