% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Riemannian_utils.R
\name{variable_selection}
\alias{variable_selection}
\title{Perform SVD-weighted BH selection on tangent coordinates.}
\usage{
variable_selection(S, fdr_level = 0.05)
}
\arguments{
\item{S}{Tangent feature matrix (d x n).}

\item{fdr_level}{Benjamini-Hochberg threshold.}
}
\value{
List with filtered \code{S} and selected feature \code{indices}.
}
\description{
Perform SVD-weighted BH selection on tangent coordinates.
}
