% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers.R
\name{fisher_lda}
\alias{fisher_lda}
\title{Compute Fisher LDA direction for binary classes.}
\usage{
fisher_lda(x, label)
}
\arguments{
\item{x}{Feature matrix with shape \verb{p x n}.}

\item{label}{Binary label vector of length \code{n}.}
}
\value{
List with projection \code{direction} and normalized sample \code{score}.
}
\description{
Compute Fisher LDA direction for binary classes.
}
