% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bandpass_Filtering.R
\name{freqBank}
\alias{freqBank}
\title{Apply a Butterworth filter bank to one trial matrix.}
\usage{
freqBank(
  eeg,
  fs,
  order = 3,
  bands = list(c(8, 13), c(13, 18), c(18, 23), c(23, 28))
)
}
\arguments{
\item{eeg}{Trial matrix (\verb{samples x channels}).}

\item{fs}{Sampling rate in Hz.}

\item{order}{Filter order.}

\item{bands}{List of pass bands, each as \code{c(low, high)} in Hz.}
}
\value{
List of filtered trial matrices, one per band.
}
\description{
Apply a Butterworth filter bank to one trial matrix.
}
