% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Riemannian_utils.R
\name{riemannian_distance}
\alias{riemannian_distance}
\title{Compute affine-invariant Riemannian distance between two SPD matrices.}
\usage{
riemannian_distance(A, B)
}
\arguments{
\item{A}{First SPD matrix.}

\item{B}{Second SPD matrix.}
}
\value{
Scalar geodesic distance.
}
\description{
Compute affine-invariant Riemannian distance between two SPD matrices.
}
